Consolidate C and C++ LSP adapters
