+++ crates/zed/src/languages.rs
@@ -4,7 +4,6 @@ use rust_embed::RustEmbed;
 use std::{borrow::Cow, str, sync::Arc};
 
 mod c;
-mod cpp;
 mod installation;
 mod json;
 mod rust;
@@ -26,7 +25,7 @@ pub fn build_language_registry(login_shell_env_loaded: Task<()>) -> LanguageRegi
         (
             "cpp",
             tree_sitter_cpp::language(),
-            Some(Arc::new(cpp::CppLspAdapter) as Arc<dyn LspAdapter>),
+            Some(Arc::new(c::CLspAdapter) as Arc<dyn LspAdapter>),
         ),
         (
             "json",
